[#176314517] refactoring
