+++ ts/sagas/wallet.ts
@@ -162,7 +162,6 @@ import { isTestEnv } from "../utils/environment";
 import { defaultRetryingFetch } from "../utils/fetch";
 import { getCurrentRouteKey, getCurrentRouteName } from "../utils/navigation";
 import { getTitleFromCard } from "../utils/paymentMethod";
-import { backoffWait } from "../utils/saga";
 import { SessionManager } from "../utils/SessionManager";
 import { hasFunctionEnabled } from "../utils/walletv2";
 import { paymentsDeleteUncompletedSaga } from "./payments";
@@ -186,6 +185,7 @@ import {
 } from "./wallet/pagopaApis";
 import { paymentIdSelector } from "../store/reducers/wallet/payment";
 import { sendAddCobadgeMessageSaga } from "./wallet/cobadgeReminder";
+import { waitBackoffError } from "../utils/backoffError";
 
 const successScreenDelay = 2000 as Millisecond;
 /**
@@ -639,7 +639,7 @@ export function* watchWalletSaga(
   yield takeLatest(getType(fetchTransactionsRequestWithExpBackoff), function* (
     action: ActionType<typeof fetchTransactionsRequestWithExpBackoff>
   ) {
-    yield call(backoffWait, fetchTransactionsFailure);
+    yield call(waitBackoffError, fetchTransactionsFailure);
     yield put(fetchTransactionsRequest(action.payload));
   });
 
@@ -673,7 +673,7 @@ export function* watchWalletSaga(
   );
 
   yield takeLatest(getType(fetchWalletsRequestWithExpBackoff), function* () {
-    yield call(backoffWait, fetchWalletsFailure);
+    yield call(waitBackoffError, fetchWalletsFailure);
     yield put(fetchWalletsRequest());
   });
 
@@ -696,7 +696,7 @@ export function* watchWalletSaga(
     function* (
       action: ActionType<typeof addWalletCreditCardWithBackoffRetryRequest>
     ) {
-      yield call(backoffWait, addWalletCreditCardFailure);
+      yield call(waitBackoffError, addWalletCreditCardFailure);
       yield put(addWalletCreditCardRequest(action.payload));
     }
   );
