+++ ts/store/reducers/wallet/index.ts
@@ -17,9 +17,6 @@ import paymentReducer, { PaymentState } from "./payment";
 import pspsByIdReducer, { PspStateById } from "./pspsById";
 import transactionsReducer, { TransactionsState } from "./transactions";
 import walletsReducer, { PersistedWalletsState, WalletsState } from "./wallets";
-import lastRequestErrorReducer, {
-  LastRequestErrorState
-} from "./lastRequestError";
 import outcomeCodeReducer, { OutcomeCodeState } from "./outcomeCode";
 
 export type WalletState = Readonly<{
@@ -31,7 +28,6 @@ export type WalletState = Readonly<{
   abi: AbiState;
   // section used for the onboarding of a new payment method. Each payment have a sub-section
   onboarding: PaymentMethodOnboardingState;
-  lastRequestError: LastRequestErrorState;
   // Section used to know the outcome of the last payment, used both when the user pay or add a credit card.
   lastPaymentOutcomeCode: OutcomeCodeState;
 }>;
@@ -54,7 +50,6 @@ const reducer = combineReducers<WalletState, Action>({
   pspsById: pspsByIdReducer,
   abi: abiReducer,
   onboarding: onboardingReducer,
-  lastRequestError: lastRequestErrorReducer,
   lastPaymentOutcomeCode: outcomeCodeReducer
 });
